// src/main/frontend/src/services/openapi/models/SubSpeciesDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DescriptionDTO } from './DescriptionDTO';
import type { EntryDTO } from './EntryDTO';
import type { ImageDTO } from './ImageDTO';
export type SubSpeciesDTO = {
    subSpecies?: EntryDTO;
    species?: EntryDTO;
    descriptions?: Array<DescriptionDTO>;
    images?: Array<ImageDTO>;
    regions?: Array<EntryDTO>;
};